Revert "Let all chunks have the same size"
This reverts commit e970c5f377b444340cdf24d1e1fde8a410e70ea9.

This solution is flawed. I cannot fill a chunk with random data, as this
random data must be known in a reference so that I can use the same
random data on a new call of "create" to produce the same output (given
that the file has not changed). So I have several choices:

* Use zero bytes instead of random bytes (is this safe?)
* Store the random data in a ChunkReference (which bloats the backup
even more)
* Just revert the change

I'll go for option #3.