Simplify error handling when restoring data
I have changed the RestoreSnapshot to not continue on any error. This
may seem like a bad idea, but I'm not sure how I'd like to handle this
in an understandable manner.

* The previous approach using "ok" return value combined with two
different places where we catch, report and throw an Exception seems
confusing
* Using an AggregateException would mean that I need to implement a
custom "ChunkyardRestoreException" which contains the ContentReference
where the problem occurred
* Passing a "Func<ContentReference, Exception>" parameter into
"RestoreSnapshot" seems the most reasonable approach, but I still don't
like it that much