Redesign and simplify ListUris
The SnapshotStore.ListUris would ignore problems when a Snapshot cannot
be restored. I thought that this would be a good idea because it would
allow commands such as GarbageCollect to continue without failure. But
now I believe that we should abort a ListUris call if we cannot load a
snapshot, as this means that there is a data corruption issue within the
repository.