Redesign SnapshotStore and ContentStore
This commit removes most of the console printing operations of Chunkyard
as well as the IContentStore interface (which was introduced as a
workaround). These printing operations will be added in a later commit.

I have redesigned SnapshotStore.cs so that it can handle all operations
found in Commands.cs. Password prompts are now implemented in a lazy
loading fashion, which means that a password prompt will only occur if
SnapshotStore.cs performs an action for which it needs to create a key.